bottomNaigator typescript적용
